Calculate membership score of one observation for each medoid

Function to calculate a membership score for one observation for each medoid based on the distance of this observation to all medoids

Usage

calculate_memb_score(dist_med, m = 2)

Arguments

dist_med

Vector of distances to medoids

m

Fuzziness exponent, which has to be a numeric of minimum 1. Defaults to 2.

Value

List with membership scores for one observation