Prediction of cluster assignments

Function to predict cluster assignments

Usage

# S3 method for class 'fuzzyclara'
predict(object, newdata, ...)

Arguments

object

an object of class "fuzzyclara"

newdata

data.frame to look variables with which to predict

...

further arguments for predict functions

Value

clustering plot tibble

Examples

# Split data intp test and training data
train_indices <- sample(x = nrow(USArrests), size = 0.7*nrow(USArrests))
USArrests_train <- USArrests[train_indices, ]
USArrests_test <- USArrests[-train_indices, ]

# Determine fuzzy clusters on training data

cc_fuzzy <- fuzzyclara(data        = USArrests_train,
                       clusters    = 3,
                       metric      = "euclidean",
                       samples     = 1,
                       sample_size = NULL,
                       type        = "fuzzy",
                       m           = 2,
                       seed        = 3526,
                       verbose     = 0)
#> Warning: The specified sample size is equal to the number of
#>     observations in the data. PAM clustering is performed on the entire data.
cc_fuzzy
#> Clustering results
#> 
#> Medoids
#> [1] "Indiana"  "New York" "Montana" 
#> 
#> Clustering
#>  [1] 3 2 3 1 2 3 2 1 1 2 1 1 2 2 2 1 1 1 1 1 2 1 3 3 2 1 3 2 3 3 1 2 3 3 3
#> 
#> Minimum average weighted distance
#> [1] 1.789263
#> 
#> Membership scores
#>                 Cluster1  Cluster2  Cluster3
#> Oklahoma       0.2630337 0.1448693 0.5920970
#> Michigan       0.2014929 0.5427660 0.2557412
#> Hawaii         0.3153240 0.2336347 0.4510413
#> Nebraska       0.4818152 0.1154408 0.4027440
#> Alaska         0.3133491 0.3471720 0.3394789
#> Rhode Island   0.3239389 0.3034310 0.3726301
#> Colorado       0.2413297 0.4320872 0.3265830
#> Vermont        0.4867983 0.1825094 0.3306923
#> Wyoming        0.4420239 0.1395860 0.4183901
#> Nevada         0.2317236 0.4839071 0.2843693
#> New Hampshire  0.4920440 0.1611722 0.3467838
#> Arkansas       0.4259825 0.1996859 0.3743316
#> California     0.2066606 0.5355144 0.2578250
#> New Mexico     0.2131976 0.5256900 0.2611124
#> Florida        0.2068770 0.5506103 0.2425127
#> Maine          0.5100066 0.1613600 0.3286333
#> South Dakota   0.5838831 0.1279757 0.2881411
#> Iowa           0.4929520 0.1560745 0.3509735
#> North Dakota   0.4805175 0.1820484 0.3374340
#> West Virginia  0.5419153 0.1520779 0.3060068
#> South Carolina 0.3179343 0.3538622 0.3282034
#> Kentucky       0.4713181 0.1476279 0.3810539
#> Tennessee      0.2964942 0.3397406 0.3637651
#> Missouri       0.2526691 0.3286269 0.4187040
#> North Carolina 0.3293921 0.3481687 0.3224392
#> Idaho          0.5436787 0.1359001 0.3204212
#> Ohio           0.2422328 0.1874984 0.5702688
#> Maryland       0.2301054 0.4974539 0.2724406
#> Connecticut    0.3661149 0.2176768 0.4162084
#> Oregon         0.3119137 0.2259963 0.4620901
#> Montana        1.0000000 0.0000000 0.0000000
#> New York       0.0000000 1.0000000 0.0000000
#> Indiana        0.0000000 0.0000000 1.0000000
#> Massachusetts  0.3013506 0.2891735 0.4094759
#> Utah           0.3128735 0.2444537 0.4426728

# Determine distance matrix of the observations and cluster medoids
USArrests_medoids <- USArrests[rownames(USArrests) %in% cc_fuzzy$medoids,]
dist <- proxy::dist(x = USArrests_test[, -1], y = USArrests_medoids[, -1],
                    method = "euclidean") #use same metric as above

# Make cluster prediction for test data

USArrests_clusters_predicted <- predict(object = cc_fuzzy,
                                        newdata = USArrests_test,
                                        dist_matrix = dist)
USArrests_clusters_predicted$membership_scores
#>                Cluster1  Cluster2  Cluster3
#> Alabama      0.29923368 0.3630613 0.3377050
#> Arizona      0.19259828 0.5661237 0.2412780
#> Delaware     0.30966984 0.3210244 0.3693058
#> Georgia      0.28502946 0.3901573 0.3248133
#> Illinois     0.09184771 0.7853221 0.1228302
#> Kansas       0.29391169 0.0996555 0.6064328
#> Louisiana    0.25522746 0.4460923 0.2986802
#> Minnesota    0.43345453 0.1605920 0.4059534
#> Mississippi  0.33434524 0.3328186 0.3328361
#> New Jersey   0.24998299 0.3820818 0.3679353
#> Pennsylvania 0.32155651 0.1597925 0.5186510
#> Texas        0.17432795 0.5843937 0.2412783
#> Virginia     0.29357345 0.1614680 0.5449586
#> Washington   0.30641477 0.2270731 0.4665121
#> Wisconsin    0.43879954 0.1733572 0.3878432