Prediction of cluster assignments

Function to predict cluster assignments

Usage

# S3 method for fuzzyclara
predict(object, newdata, ...)

Arguments

object

an object of class "fuzzyclara"

newdata

data.frame to look variables with which to predict

...

further arguments for predict functions

Value

clustering plot tibble

Examples

# Split data intp test and training data
train_indices <- sample(x = nrow(USArrests), size = 0.7*nrow(USArrests))
USArrests_train <- USArrests[train_indices, ]
USArrests_test <- USArrests[-train_indices, ]

# Determine fuzzy clusters on training data

cc_fuzzy <- fuzzyclara(data        = USArrests_train,
                       clusters    = 3,
                       metric      = "euclidean",
                       samples     = 1,
                       sample_size = NULL,
                       type        = "fuzzy",
                       m           = 2,
                       seed        = 3526,
                       verbose     = 0)
#> Warning: The specified sample size is equal to the number of
#>     observations in the data. PAM clustering is performed on the entire data.
cc_fuzzy
#> Clustering results
#> 
#> Medoids
#> [1] "Illinois"   "Iowa"       "New Mexico"
#> 
#> Clustering
#>  [1] 2 1 3 1 2 1 1 1 2 1 1 2 1 1 3 3 1 1 1 2 3 1 3 1 1 1 3 3 3 3 3 1 2 2 3
#> 
#> Minimum average weighted distance
#> [1] 1.820894
#> 
#> Membership scores
#>                  Cluster1   Cluster2   Cluster3
#> New Jersey     0.24718861 0.48237598 0.27043541
#> Iowa           1.00000000 0.00000000 0.00000000
#> Mississippi    0.24790256 0.35132408 0.40077335
#> Minnesota      0.69905029 0.16368809 0.13726162
#> Texas          0.12345801 0.55318921 0.32335277
#> Connecticut    0.51583355 0.27547030 0.20869614
#> New Hampshire  0.90804563 0.04895042 0.04300396
#> Wisconsin      0.75963117 0.13085791 0.10951092
#> Washington     0.33823414 0.36186254 0.29990333
#> Montana        0.54578072 0.23778453 0.21643475
#> Pennsylvania   0.45812427 0.30962265 0.23225308
#> New York       0.06379656 0.73669808 0.19950536
#> Vermont        0.59538846 0.20541076 0.19920078
#> Nebraska       0.60374130 0.21430863 0.18195007
#> California     0.15657041 0.40224055 0.44118904
#> Colorado       0.17458473 0.38122664 0.44418863
#> Kansas         0.46255911 0.29734968 0.24009121
#> West Virginia  0.57966261 0.21585295 0.20448444
#> Rhode Island   0.37995542 0.36412673 0.25591785
#> Massachusetts  0.35028903 0.38885124 0.26085973
#> Tennessee      0.19438419 0.36134659 0.44426922
#> Hawaii         0.42395877 0.32526539 0.25077584
#> Arkansas       0.32102966 0.33440928 0.34456106
#> Idaho          0.66834137 0.17477339 0.15688525
#> Indiana        0.39504944 0.32964725 0.27530330
#> Maine          0.76538036 0.12389408 0.11072556
#> South Carolina 0.20710444 0.34285124 0.45004432
#> Michigan       0.08521270 0.26593028 0.64885701
#> Florida        0.12470080 0.36761731 0.50768188
#> North Carolina 0.23650602 0.35269274 0.41080124
#> Alaska         0.20572372 0.30709545 0.48718083
#> North Dakota   0.69990429 0.15615110 0.14394461
#> Illinois       0.00000000 1.00000000 0.00000000
#> Virginia       0.30722136 0.37496615 0.31781249
#> New Mexico     0.00000000 0.00000000 1.00000000

# Determine distance matrix of the observations and cluster medoids
USArrests_medoids <- USArrests[rownames(USArrests) %in% cc_fuzzy$medoids,]
dist <- proxy::dist(x = USArrests_test[, -1], y = USArrests_medoids[, -1],
                    method = "euclidean") #use same metric as above

# Make cluster prediction for test data

USArrests_clusters_predicted <- predict(object = cc_fuzzy, newdata = USArrests_test, dist_matrix = dist)
USArrests_clusters_predicted$membership_scores
#>                Cluster1  Cluster2  Cluster3
#> Alabama      0.19191979 0.3787156 0.4293647
#> Arizona      0.11519594 0.4308609 0.4539431
#> Delaware     0.26512774 0.4339358 0.3009365
#> Georgia      0.19378028 0.3673066 0.4389131
#> Kentucky     0.41628777 0.3027335 0.2809787
#> Louisiana    0.16101069 0.4027210 0.4362683
#> Maryland     0.08475208 0.2674779 0.6477701
#> Missouri     0.18943105 0.4244784 0.3860905
#> Nevada       0.16050478 0.3490638 0.4904314
#> Ohio         0.32681174 0.3910569 0.2821313
#> Oklahoma     0.34610772 0.3687428 0.2851495
#> Oregon       0.30094477 0.3588833 0.3401719
#> South Dakota 0.66792602 0.1716479 0.1604261
#> Utah         0.38258554 0.3492397 0.2681748
#> Wyoming      0.40974181 0.3207744 0.2694838